frontend/pyproject.toml
[project]
name = "chatai-frontend"
version = "1.0.0"
description = "FastAPI Chat AI Test Interface"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.5.0",
    "jinja2>=3.1.0",
]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "httpx>=0.26.0",
]
